requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
asyncpg
python-jose
argon2-cffi
pydantic[email]